+++ src/main/java/org/octri/notification/batch/NotificationBatchJob.java
@@ -8,12 +8,10 @@ import org.springframework.batch.core.configuration.DuplicateJobException;
 import org.springframework.batch.core.explore.JobExplorer;
 import org.springframework.batch.core.launch.JobLauncher;
 import org.springframework.scheduling.annotation.Scheduled;
-import org.springframework.stereotype.Component;
 
 /**
  * The component that handles running the job if it's not already in progress
  */
-@Component
 public class NotificationBatchJob {
 
 	private final JobExplorer jobExplorer;
@@ -70,7 +68,7 @@ public class NotificationBatchJob {
 	 * @throws Exception
 	 *             exception thrown by the jobLauncher
 	 */
-	@Scheduled(cron = "${octri.notifications.schedule:@yearly}")
+	@Scheduled(cron = "${octri.notifications.schedule:@yearly}", scheduler = "notificationTaskScheduler")
 	public void runScheduledNotificationJob() throws Exception {
 		runNotificationJob(ProcessingMode.SCHEDULED);
 	}