+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-- Initial implementation of the project
+### Added
+
+- Initial implementation of the library (CIS-3203)
+- Add a dedicated Task Scheduler for the batch jobs (CIS-3347)
+- README documentation (CIS-3349)
+
+### Changed
+
+- Use Spring Boot Autoconfiguration (CIS-3347)
 
 [unreleased]: https://github.com/OHSU-OCTRI/notification-lib/